Refactor: rename `models` to `schemas` for clarity.
