+++ src/endfield_essence_recognizer/schemas/__init__.py
@@ -1,0 +1,7 @@
+"""
+Stores schema definitions used in FastAPI endpoints.
+
+Contains Enums, Pydantic models, etc. that define the
+interface for API requests and responses. Enums may
+also be used in other parts of the application.
+"""
