+++ src/endfield_essence_recognizer/schemas/scanner.py
