+++ src/endfield_essence_recognizer/schemas/screenshot.py
