+++ src/endfield_essence_recognizer/services/screenshot_service.py
@@ -8,7 +8,7 @@ from endfield_essence_recognizer.core.layout.base import (
     ResolutionProfile,
 )
 from endfield_essence_recognizer.core.window import WindowManager
-from endfield_essence_recognizer.models.screenshot import (
+from endfield_essence_recognizer.schemas.screenshot import (
     ImageFormat,
     ScreenshotSaveFormat,
 )
