+++ tests/unit/core/scanner/test_action_logic.py
@@ -13,7 +13,7 @@ from endfield_essence_recognizer.core.scanner.models import (
     EssenceQuality,
     EvaluationResult,
 )
-from endfield_essence_recognizer.models.user_setting import Action, UserSetting
+from endfield_essence_recognizer.schemas.user_setting import Action, UserSetting
 
 
 @pytest.fixture
